"Update tests for creating Saisie and Project; add dynamically generated data for demography section"
This commit updates the tests for creating a Saisie and a Project in the web application. It includes logging in, navigating to the appropriate pages, and entering required information. The commit also adds a function to add dynamically generated data for the demography section of the Project.

The changes include:

* Updating the login credentials for the B300NQA user
* Adding a check to ensure that the login is successful before proceeding
* Adding a function to create a Saisie, including adding an établissement, apporteur, and project description
* Adding a function to create a Project, including assigning the Project and opening it
* Adding a function to add dynamically generated data for the demography section of the Project

The changes also include updates to the data used for the établissement and project description, as well as updates to the selectors used to interact with the application.