Update 'Contrat-Santé-SM-K.cy.js' to include a test for creating a project
The 'Contrat-Santé-SM-K.cy.js' file has been updated to include a new test for creating a project. The test logs in, creates a request, and then creates a project by adding necessary information such as the establishment, apporteur, project description, and dynamic data.

Additionally, a new 'demography.json' file has been added to the 'data-json' folder. This file contains data used in the new project creation test.

The 'demography' object in the 'demography.json' file has been updated to include a new 'legalForm' object, which contains information about the legal form of the establishment. This information is used in the new project creation test.

Note: There are some inconsistencies in the use of single and double quotes in the code. It is recommended to use a consistent style throughout the codebase.